+++ hyper_resilient_experiments/alexnet_tune.py
@@ -50,13 +50,13 @@ def model_attack(model, model_type, attack_type, config, num_classes=NUM_CLASSES
             f = open('/lus/theta-fs0/projects/CVD-Mol-AI/mzvyagin/alexnet_datasets/fashion_splits.pkl', 'rb')
             data = pickle.load(f)
             (x_train, y_train), (x_val, y_val), (x_test, y_test) = data
-            data = DataLoader(Fashion_NP_Dataset(x_test, y_test),
+            data = DataLoader(Fashion_NP_Dataset(x_test.astype(np.float32), y_test.astype(np.float32)),
                           batch_size=int(config['batch_size']), shuffle=False)
         elif MODEL_TYPE == "caltech":
             f = open('/lus/theta-fs0/projects/CVD-Mol-AI/mzvyagin/alexnet_datasets/caltech_splits.pkl', 'rb')
             data = pickle.load(f)
             (x_train, y_train), (x_val, y_val), (x_test, y_test) = data
-            data = DataLoader(Caltech_NP_Dataset(x_test, y_test),
+            data = DataLoader(Caltech_NP_Dataset(x_test.astype(np.float32), y_test.astype(np.float32)),
                               batch_size=int(config['batch_size']), shuffle=False)
         else:
             print("NOT IMPLEMENTED")
@@ -72,12 +72,12 @@ def model_attack(model, model_type, attack_type, config, num_classes=NUM_CLASSES
             f = open('/lus/theta-fs0/projects/CVD-Mol-AI/mzvyagin/alexnet_datasets/fashion_splits.pkl', 'rb')
             data = pickle.load(f)
             (x_train, y_train), (x_val, y_val), (x_test, y_test) = data
-            data = tf.data.Dataset.from_tensor_slices((x_test, y_test)).batch(config['batch_size'])
+            data = tf.data.Dataset.from_tensor_slices((x_test.astype(np.float32), y_test.astype(np.float32))).batch(config['batch_size'])
         elif MODEL_TYPE == "caltech":
             f = open('/lus/theta-fs0/projects/CVD-Mol-AI/mzvyagin/alexnet_datasets/caltech_splits.pkl', 'rb')
             data = pickle.load(f)
             (x_train, y_train), (x_val, y_val), (x_test, y_test) = data
-            data = tf.data.Dataset.from_tensor_slices((x_test, y_test)).batch(config['batch_size'])
+            data = tf.data.Dataset.from_tensor_slices((x_test.astype(np.float32), y_test.astype(np.float32))).batch(config['batch_size'])
         else:
             print("NOT IMPLEMENTED")
             sys.exit()
